vsphere: Refactor code for common CAA framwork
Fixes #581

Also extended network wait timeout and cleaned up logging.
